Move externally-used ast utils to @kubb/kit, rename collect/collectLazy
@kubb/ast keeps only the ref and schema-graph helpers its own node
builders depend on (resolveRefName, findCircularSchemas,
collectUsedSchemaNames). childName, enumPropName, extractRefName,
isStringType, mergeAdjacentObjectsLazy, syncSchemaRef, and
containsCircularRef move to @kubb/kit, reached through kubb/kit like
the macro presets. @kubb/adapter-oas picks these up from @kubb/kit.

Also renames the visitor collection passes so the names match their
behavior: the lazy generator collectLazy is now collect, and the
eager array form (previously collect) is now collectSync. Exposes
collect (the renamed collectLazy) as a public @kubb/ast export so
@kubb/kit's containsCircularRef can keep its early-exit behavior.
