+++ packages/adapter-oas/src/converters.ts
@@ -1,5 +1,5 @@
-import { ast, childName, enumPropName, extractRefName, mergeAdjacentObjectsLazy } from '@kubb/ast'
-import { macroDiscriminatorEnum, macroEnumName, macroSimplifyUnion } from '@kubb/kit'
+import { ast } from '@kubb/ast'
+import { childName, enumPropName, extractRefName, macroDiscriminatorEnum, macroEnumName, macroSimplifyUnion, mergeAdjacentObjectsLazy } from '@kubb/kit'
 import { enumDescriptionKeys, enumExtensionKeys, SCHEMA_REF_PREFIX } from './constants.ts'
 import { createDiscriminantNode, findDiscriminator } from './discriminator.ts'
 import { isBinary, isDiscriminator, isNullable, isReference } from './oas.ts'
