+++ packages/ast/src/utils/index.ts
@@ -1,5 +1,4 @@
 export { combineExports, combineImports, combineSources } from './combineFileMembers.ts'
 export { extractStringsFromNodes } from './extractStringsFromNodes.ts'
-export { mergeAdjacentObjectsLazy } from './mergeAdjacentSchemas.ts'
-export { childName, enumPropName, extractRefName, isStringType, resolveRefName, syncSchemaRef } from './refs.ts'
-export { collectUsedSchemaNames, containsCircularRef, findCircularSchemas } from './schemaGraph.ts'
+export { resolveRefName } from './refs.ts'
+export { collectUsedSchemaNames, findCircularSchemas } from './schemaGraph.ts'
