+++ packages/kit/src/index.ts
@@ -1,6 +1,7 @@
 export { ast } from '@kubb/ast'
 export { Url } from '@internals/utils'
 export { macroDiscriminatorEnum, macroEnumName, macroRenameSchema, macroSimplifyUnion } from './macros/index.ts'
+export { childName, containsCircularRef, enumPropName, extractRefName, isStringType, mergeAdjacentObjectsLazy, syncSchemaRef } from './utils/index.ts'
 export { createAdapter } from '@kubb/core'
 export { createRenderer } from '@kubb/core'
 export { createStorage } from '@kubb/core'
