+++ packages/kit/src/utils/index.ts
@@ -1,0 +1,3 @@
+export { mergeAdjacentObjectsLazy } from './mergeAdjacentSchemas.ts'
+export { childName, enumPropName, extractRefName, isStringType, syncSchemaRef } from './refs.ts'
+export { containsCircularRef } from './schemaGraph.ts'
